Specify entry point for crawler / processor
And drop `autoload`.
